Implement DB call/processing

+++ api/src/main/java/com/codeforcommunity/dto/map/SiteFeatureProperties.java
@@ -1,12 +1,13 @@
 package com.codeforcommunity.dto.map;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 public class SiteFeatureProperties {
   Integer id;
   Boolean treePresent;
-  Double diameter;
+  BigDecimal diameter;
   String species;
   Timestamp updatedAt;
   String updatedBy; // username
@@ -15,7 +16,7 @@ public class SiteFeatureProperties {
   public SiteFeatureProperties(
       Integer id,
       Boolean treePresent,
-      Double diameter,
+      BigDecimal diameter,
       String species,
       Timestamp updatedAt,
       String updatedBy,
@@ -38,7 +39,7 @@ public class SiteFeatureProperties {
     return treePresent;
   }
 
-  public Double getDiameter() {
+  public BigDecimal getDiameter() {
     return diameter;
   }
